api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
pydantic==2.10.4
PyJWT==2.10.1
httpx==0.28.1
aioapns==3.3
firebase-admin==6.6.0
